Relocate Pro welcome web presenter to clear file-length incidental gate
The prior fix raised PricingPlansScreen.swift's budget to 893, but the
guard's incidental-growth gate fails any tracked (>=500 line) file whose
PR growth exceeds 25 lines regardless of the checked-in budget, so the
merge stayed red at +46. Move presentProWelcomeWeb and decoratedAppWebURL
into an extension in the already-wired ProWelcomeChecklist.swift, making
presentBrowserSplit internal so the extension can reach it. PricingPlansScreen
drops to 858 (+11 over base, within the 25 allowance); budget entry set to 858.
